requirements.txt
python-dotenv
flask
openai==1.55.3
requests
streamlit
beautifulsoup4
langchain==0.3.24

